changing to double

+++ test/slogo/model/movementcommand/SetPositionTest.java
@@ -26,7 +26,7 @@ public class SetPositionTest {
   @Test
   void moveToNewLocation() {
     setPosition = new SetPosition(new Constant(10), new Constant(15));
-    assertEquals(18, setPosition.execute(turtle));
+    assertEquals(18, (int) setPosition.execute(turtle));
     assertEquals(10, turtle.getXCoordinate());
     assertEquals(15, turtle.getYCoordinate());
   }
